wiki-organizer/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
httpx==0.28.1
python-multipart==0.0.17
python-docx==1.1.2
pypdf==5.1.0
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.6.1
tenacity==9.0.0
